Fix chunk size for big configurations
XENSTORE_PAYLOAD_MAX is the maximum message size that can go into the
xenstore protocol, therefore for a `xs_write` it includes the path and a
null character for separator, so we can't use this macro directly for
chunk size.

We could calculate the chunk size based on the path length, however
ClickOS is not accepting big chunk sizes for some reason. So for now set
the chunk size for a fixed value of 512, that is known to work.

This fixes #3.